third_party/WebKit/LayoutTests/css3/calc/table-border-spacing-expected.txt

PASS document.getElementById("stimulus").offsetWidth is 416
PASS document.getElementById("stimulus").offsetHeight is 205
PASS successfullyParsed is true

TEST COMPLETE

				
				
